pyproject.toml
[project]
name = "swearjar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "better-profanity>=0.7.0",
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.0",
]
